Move ComponentElementsTag to capability package
