+++ subprojects/platform-base/src/main/java/dev/nokee/platform/base/internal/ComponentTasksPropertyRegistrationFactory.java
@@ -27,6 +27,7 @@ import dev.nokee.model.internal.state.ModelState;
 import dev.nokee.model.internal.state.ModelStates;
 import dev.nokee.platform.base.TaskView;
 import dev.nokee.platform.base.internal.elements.ComponentElementTypeComponent;
+import dev.nokee.platform.base.internal.elements.ComponentElementsTag;
 import lombok.val;
 import org.apache.commons.lang3.StringUtils;
 import org.gradle.api.Task;
