+++ subprojects/platform-base/src/main/java/dev/nokee/platform/base/internal/elements/ComponentElementsTag.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.platform.base.internal;
+package dev.nokee.platform.base.internal.elements;
 
 import dev.nokee.model.internal.core.ModelComponent;
